Convert project to TS. Added eslint/prettier
