--- src/App.js
@@ -1,13 +1,0 @@
-import * as React from "react";
-import { Admin, Resource, ListGuesser } from 'react-admin';
-import { UserList } from './users';
-import jsonServerProvider from 'ra-data-json-server';
-
-const dataProvider = jsonServerProvider('https://jsonplaceholder.typicode.com');
-const App = () => (
-  <Admin dataProvider={dataProvider}>
-    <Resource name="users" list={UserList}/>
-  </Admin>
-);
-
-export default App;
