+++ src/App.tsx
@@ -1,0 +1,15 @@
+import * as React from "react";
+import { Admin, Resource } from "react-admin";
+import jsonServerProvider from "ra-data-json-server";
+import { UserList } from "./users";
+
+import "./App.css";
+
+const dataProvider = jsonServerProvider("https://jsonplaceholder.typicode.com");
+const App = () => (
+  <Admin dataProvider={dataProvider}>
+    <Resource name="users" list={UserList} />
+  </Admin>
+);
+
+export default App;
