+++ src/users.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+import { FC } from "react";
+import { List, Datagrid, TextField, EmailField, UrlField, ListProps } from "react-admin";
+
+export const UserList: FC<ListProps> = (props) => (
+  <List {...props}>
+    <Datagrid rowClick="edit">
+      <TextField source="id" />
+      <TextField source="name" />
+      <EmailField source="email" />
+      <TextField source="phone" />
+      <UrlField source="website" />
+      <TextField source="company.name" />
+    </Datagrid>
+  </List>
+);
